Add NoSelectTable component and refactor ScoreTable to use it
The UI components have been expanded by adding the NoSelectTable component, which is a table that disables user selection. Also, ScoreTable, a component of the scoring system, has been adapted to use NoSelectTable instead of the previous table component. This refactoring allows better control and customization in terms of user interactions.